move camera to main
