+++ ProjektExpanded/main/camera_stream.h
@@ -2,12 +2,18 @@
 #define CAMERA_STREAM_H
 
 #include "esp_err.h"
+#include "esp_camera.h"
 #include <stdbool.h>
 #include <stdint.h>
 #include <stddef.h>
 
+// Global variables for camera streaming
+extern camera_fb_t *current_frame;
+extern int sock;
+extern bool socket_connected;
+
 esp_err_t init_camera(void);
-void stream_camera_task(void *pvParameters);
+// stream_camera_task moved to main.c
 bool send_frame_over_websocket(const uint8_t *frame_data, size_t frame_len);
 bool connect_socket(void);
